Respond to review comments by @rphmeier
